separate core from models, update module resolution
