+++ src/browser.ts
@@ -1,3 +1,4 @@
 export * as Models from './models/index.js';
 export * as Types from './types.js';
-export * as Views from './views/index.js';
+export * from './views/index.js';
+export * from './core/index.js';
