+++ src/core/dashboard.ts
