+++ src/index.ts
@@ -3,3 +3,4 @@ export * as Models from './models/index.js';
 export * as Types from './types.js';
 export * as Views from './views/index.js';
 export * from './tinystacks-error.js';
+export * from './core/index.js';
