+++ src/models/base-provider.ts
@@ -1,5 +1,5 @@
 import { Provider as ProviderType } from '@tinystacks/ops-model';
-import { Provider } from './provider.js';
+import { Provider } from '../core/provider.js';
 
 export class BaseProvider extends Provider {
   static fromJson (props: ProviderType): BaseProvider {
