+++ src/models/index.ts
@@ -1,7 +1,3 @@
 export * from './base-provider.js';
 export * from './base-widget.js';
-export * from './console.js';
-export * from './dashboard.js';
-export * from './parsable.js';
-export * from './provider.js';
 export * from './widget.js';
