+++ src/node.ts
@@ -1,3 +1,4 @@
-export * as Controllers from './controllers/index.js';
 export * as Models from './models/index.js';
 export * as Types from './types.js';
+export * from './controllers/index.js';
+export * from './core/index.js';
